App.css - added the rest of the css for the application; viz. game, lost player, won player
App.tsx - tested the display of the game. It doesn't work, but that is because the game thus far has been implemented using direct handling of the DOM, which should rarely if ever be done in React.

index.css - removed all of the logic here to test if it would make the actors viewable. Upon testing, the actors were not visible because of the relative position of the game.

vite-env.d.ts - removed all of the contents here to keep everything more modular.

gameLogic.ts -
Moved the Level class into it's own file. Level.ts
Implemented the DOMDisplay class.

Actors.d.ts -
the data types of Actors.

Settings.ts - moved Scale into it's own file.